Changed dicemal typr to float

+++ database/migrations/2024_08_20_112210_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->text('description');
             $table->string("sku");
-            $table->decimal('price', 10, 2);
+            $table->float('price');
             $table->integer("quantity");
             $table->foreignId('category_id')->constrained('categories')->onDelete('cascade');
             $table->foreignId('supplier_id')->constrained('suppliers')->onDelete('cascade');
